added macro for const and main inputs sections
